Support Raspberry Pi Zero
On 32-bit Linux, several kernel-derived structs use narrower seconds-like
fields than on 64-bit (Timex/Timespec fields are int32; C time_t is 32-bit
in the ntp SHM struct). The code assumed everything was int64.

- phc, ntptime: introduce kernelLong alias (int64 on amd64/arm64, int32
  on arm) to convert at the boundary between Timex/Timespec and PTP
  ioctl structs, which always use int64 seconds.
- ntpshm: split ztypes_linux.go per-arch so the cgo-emitted struct
  matches each platform's C layout. Add a shmSec alias for the seconds
  fields and derive expectedSize from C.sizeof_struct_shmTime so it
  travels with the layout.
- devnotify: type the udev magic constant as uint32 so 0xfeedcafe
  doesn't overflow the default int type on 32-bit.
- daemon, promobs: widen ClockAccuracy field and matching parameter
  from int to int64; valid PTP accuracies extend to 10 s.
- Tests: avoid int32 overflow in time-arithmetic helpers (use
  time.Duration), and use math.MaxInt / math.MinInt / math.MaxUint
  for platform-aware integer boundary tests.

Refs #305

+++ time/app/daemon/config.go
@@ -69,7 +69,7 @@ type LeapSecondConfig struct {
 }
 
 type PTPConfig struct {
-	ClockAccuracy           int          `toml:"clockAccuracy"`
+	ClockAccuracy           int64        `toml:"clockAccuracy"`
 	OffsetScaledLogVariance uint16       `toml:"offsetScaledLogVariance"`
 	AllanDeviation          float64      `toml:"allanDeviation"`
 	DomainNumber            uint8        `toml:"domainNumber"`
